Agrego excepcion en caso de que haya error del servidor a la hora de loguearse. Hago que todos los form home hereden de formbasehome y creo metodo ConfiguraDgv para cada form home

+++ SmileSoft.Services/PacienteService.cs
@@ -30,6 +30,19 @@ namespace SmileSoft.Services
         public bool Delete(int id)
         {
             var pacienteRepository = new PacienteRepository();
+            var atencionRepository = new AtencionRepository();
+            var paciente = pacienteRepository.Get(id);
+            if (paciente == null)
+            {
+                throw new Exception("No se encontró el paciente.");
+            }
+            var atencionesDelPaciente = paciente.Atenciones;
+            // Verificar si el paciente tiene atenciones asociadas
+            if (atencionesDelPaciente != null && atencionesDelPaciente.Any())
+            {
+                throw new Exception("No se puede eliminar el paciente porque tiene atenciones asociadas.");
+            }
+            
             return pacienteRepository.Delete(id);
         }
         public PacienteDTO GetPaciente(int id)
